+++ components/collector/src/source_collectors/api_source_collectors/azure_devops.py
@@ -7,6 +7,7 @@ See https://docs.microsoft.com/en-gb/rest/api/azure/devops/?view=azure-devops-re
 from abc import ABC
 from datetime import datetime
 from typing import cast, Any, Dict, Final, List, Tuple
+import urllib.parse
 
 from dateutil.parser import parse
 import aiohttp
@@ -71,7 +72,7 @@ class AzureDevopsRepositoryBase(SourceCollector, ABC):  # pylint: disable=abstra
     async def _repository_id(self) -> str:
         """Return the repository id belonging to the repository."""
         api_url = str(await super()._api_url())
-        repository = self._parameter("repository") or api_url.rsplit("/", 1)[-1]
+        repository = self._parameter("repository") or urllib.parse.unquote(api_url.rsplit("/", 1)[-1])
         repositories_url = URL(f"{api_url}/_apis/git/repositories?api-version=4.1")
         repositories = (await (await super()._get_source_responses(repositories_url))[0].json())["value"]
         return str([r for r in repositories if repository in (r["name"], r["id"])][0]["id"])