+++ components/collector/tests/source_collectors/api_source_collectors/test_azure_devops.py
@@ -10,10 +10,12 @@ class AzureDevopsTestCase(SourceCollectorTestCase):
     """Base class for testing Azure DevOps collectors."""
     def setUp(self):
         super().setUp()
+        self.url = "https://azure_devops/org/project"
+        self.work_item_url = "https://work_item"
         self.sources = dict(
-            source_id=dict(type="azure_devops", parameters=dict(url="https://azure_devops", private_token="xxx")))
+            source_id=dict(type="azure_devops", parameters=dict(url=self.url, private_token="xxx")))
         self.work_item = dict(
-            id="id", url="https://url",
+            id="id", url=self.work_item_url,
             fields={"System.TeamProject": "Project", "System.Title": "Title", "System.WorkItemType": "Task",
                     "System.State": "New", "Microsoft.VSTS.Scheduling.StoryPoints": 2.0})
 
@@ -45,7 +47,8 @@ class AzureDevopsIssuesTest(AzureDevopsTestCase):
         self.assert_measurement(
             response,
             entities=[dict(
-                key="id", project="Project", title="Title", work_item_type="Task", state="New", url="https://url")])
+                key="id", project="Project", title="Title", work_item_type="Task", state="New",
+                url=self.work_item_url)])
 
 
 class AzureDevopsReadyStoryPointsTest(AzureDevopsTestCase):
@@ -70,24 +73,21 @@ class AzureDevopsReadyStoryPointsTest(AzureDevopsTestCase):
         self.assert_measurement(response, value="0", entities=[])
 
 
-class AzureDevopsUnmergedBranchesTest(SourceCollectorTestCase):
+class AzureDevopsUnmergedBranchesTest(AzureDevopsTestCase):
     """Unit tests for the Azure DevOps Server unmerged branches."""
 
     def setUp(self):
         super().setUp()
-        self.sources = dict(
-            source_id=dict(
-                type="azure_devops", parameters=dict(
-                    url="https://azure_devops/org/project", private_token="xxx", branches_to_ignore=["ignored_.*"])))
+        self.sources["source_id"]["parameters"]["branches_to_ignore"] = ["ignored_.*"]
         self.metric = dict(type="unmerged_branches", sources=self.sources, addition="sum")
         self.repositories = dict(value=[dict(id="id", name="project")])
+        self.landing_url = f"{self.url}/_git/project/branches"
 
     async def test_no_branches_except_master(self):
         """Test that the number of unmerged branches is returned."""
         branches = dict(value=[dict(name="master", isBaseVersion=True)])
         response = await self.collect(self.metric, get_request_json_side_effect=[self.repositories, branches])
-        self.assert_measurement(
-            response, value="0", entities=[], landing_url="https://azure_devops/org/project/_git/project/branches")
+        self.assert_measurement(response, value="0", entities=[], landing_url=self.landing_url)
 
     async def test_unmerged_branches(self):
         """Test that the number of unmerged branches is returned."""
@@ -95,49 +95,40 @@ class AzureDevopsUnmergedBranchesTest(SourceCollectorTestCase):
         branches = dict(
             value=[
                 dict(name="master", isBaseVersion=True),
-                dict(name="branch", isBaseVersion=False, aheadCount=1,
-                     commit=dict(committer=dict(date=timestamp))),
+                dict(name="branch", isBaseVersion=False, aheadCount=1, commit=dict(committer=dict(date=timestamp))),
                 dict(name="ignored_branch", isBaseVersion=False, aheadCount=1,
                      commit=dict(committer=dict(date=timestamp)))])
         response = await self.collect(self.metric, get_request_json_side_effect=[self.repositories, branches])
         expected_age = str(days_ago(parse(timestamp)))
         self.assert_measurement(
-            response, value="1",
-            entities=[dict(name="branch", key="branch", commit_age=expected_age, commit_date="2019-09-03")],
-            landing_url="https://azure_devops/org/project/_git/project/branches")
+            response, value="1", landing_url=self.landing_url,
+            entities=[dict(name="branch", key="branch", commit_age=expected_age, commit_date="2019-09-03")])
 
 
-class AzureDevopsSourceUpToDatenessTest(SourceCollectorTestCase):
+class AzureDevopsSourceUpToDatenessTest(AzureDevopsTestCase):
     """Unit tests for the Azure DevOps Server source up-to-dateness collector."""
 
     async def test_age(self):
         """Test that the age of the file is returned."""
-        sources = dict(
-            source_id=dict(
-                type="azure_devops",
-                parameters=dict(
-                    url="https://azure_devops/org/project", repository="repo", private_token="xxx",
-                    file_path="README.md")))
-        metric = dict(type="source_up_to_dateness", sources=sources, addition="max")
+        self.sources["source_id"]["parameters"]["repository"] = "repo"
+        self.sources["source_id"]["parameters"]["file_path"] = "README.md"
+        metric = dict(type="source_up_to_dateness", sources=self.sources, addition="max")
         repositories = dict(value=[dict(id="id", name="repo")])
         timestamp = "2019-09-03T20:43:00Z"
         commits = dict(value=[dict(committer=dict(date=timestamp))])
         response = await self.collect(metric, get_request_json_side_effect=[repositories, commits])
         expected_age = str(days_ago(parse(timestamp)))
         self.assert_measurement(
-            response, value=expected_age,
-            landing_url="https://azure_devops/org/project/_git/repo?path=README.md&version=GBmaster")
+            response, value=expected_age, landing_url=f"{self.url}/_git/repo?path=README.md&version=GBmaster")
 
 
-class AzureDevopsTestsTest(SourceCollectorTestCase):
+class AzureDevopsTestsTest(AzureDevopsTestCase):
     """Unit tests for the Azure DevOps Server tests collector."""
 
     async def test_nr_of_tests(self):
         """Test that the number of tests is returned."""
-        sources = dict(
-            source_id=dict(
-                type="azure_devops", parameters=dict(url="https://azure_devops", private_token="xxx", test_result=[])))
-        metric = dict(type="tests", sources=sources, addition="sum")
+        self.sources["source_id"]["test_result"] = []
+        metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(
             metric, get_request_json_return_value=dict(
                 value=[
@@ -149,73 +140,64 @@ class AzureDevopsTestsTest(SourceCollectorTestCase):
 
     async def test_nr_of_failed_tests(self):
         """Test that the number of failed tests is returned."""
-        sources = dict(
-            source_id=dict(
-                type="azure_devops",
-                parameters=dict(url="https://azure_devops", private_token="xxx", test_result=["failed"])))
-        metric = dict(type="tests", sources=sources, addition="sum")
+        self.sources["source_id"]["test_result"] = ["failed"]
+        metric = dict(type="tests", sources=self.sources, addition="sum")
         response = await self.collect(
             metric, get_request_json_return_value=dict(value=[dict(build=dict(id="1"), unanalyzedTests=4)]))
         self.assert_measurement(response, value="4")
 
 
-class AzureDevopsFailedJobsTest(SourceCollectorTestCase):
+class AzureDevopsFailedJobsTest(AzureDevopsTestCase):
     """Unit tests for the Azure Devops Server failed jobs collector."""
+
+    def setUp(self):
+        super().setUp()
+        self.path = r"\\folder"
+        self.pipeline = r"folder/pipeline"
+        self.api_url = f"{self.url}/_apis/build/definitions?includeLatestBuilds=true&api-version=4.1"
 
     async def test_nr_of_failed_jobs(self):
         """Test that the number of failed jobs is returned, that pipelines can be ignored by status, by name, and by
         regular expression."""
-        sources = dict(
-            source_id=dict(
-                type="azure_devops",
-                parameters=dict(
-                    url="https://azure_devops", private_token="xxx", failure_type=["failed"],
-                    jobs_to_ignore=["ignore_by_name", "folder/ignore.*"])))
-        metric = dict(type="failed_jobs", sources=sources, addition="sum")
+        self.sources["source_id"]["parameters"]["failure_type"] = ["failed"]
+        self.sources["source_id"]["parameters"]["jobs_to_ignore"] = ["ignore_by_name", "folder/ignore.*"]
+        metric = dict(type="failed_jobs", sources=self.sources, addition="sum")
         response = await self.collect(
             metric,
             get_request_json_return_value=dict(
                 value=[
-                    dict(path=r"\\folder", name="pipeline", _links=dict(web=dict(href="https://azure_devops/build")),
+                    dict(path=self.path, name="pipeline", _links=dict(web=dict(href=f"{self.url}/build")),
                          latestCompletedBuild=dict(result="failed", finishTime="2019-11-15T12:24:10.1905868Z")),
-                    dict(path=r"\\folder\\subfolder", name="pipeline", latestCompletedBuild=dict(result="canceled")),
-                    dict(path=r"\\folder", name="ignore_by_re", latestCompletedBuild=dict(result="failed")),
+                    dict(path=fr"{self.path}\\subfolder", name="pipeline",
+                         latestCompletedBuild=dict(result="canceled")),
+                    dict(path=self.path, name="ignore_by_re", latestCompletedBuild=dict(result="failed")),
                     dict(path=r"\\", name="ignore_by_name", latestCompletedBuild=dict(result="failed")),
                     dict(path=r"\\", name="no_builds")
                 ]))
         expected_age = days_ago(parse("2019-11-15T12:24:10.1905868Z"))
         self.assert_measurement(
-            response, value="1",
-            api_url="https://azure_devops/_apis/build/definitions?includeLatestBuilds=true&api-version=4.1",
-            landing_url="https://azure_devops/_build",
+            response, value="1", landing_url=f"{self.url}/_build", api_url=self.api_url,
             entities=[
-                dict(name=r"folder/pipeline", key=r"folder/pipeline", url="https://azure_devops/build",
+                dict(name=self.pipeline, key=self.pipeline, url=f"{self.url}/build",
                      build_date="2019-11-15", build_age=str(expected_age), build_status="failed")])
 
     async def test_nr_of_unused_jobs(self):
         """Test that the number of unused jobs is returned, that pipelines can be ignored by name and by
         regular expression."""
-        sources = dict(
-            source_id=dict(
-                type="azure_devops",
-                parameters=dict(
-                    url="https://azure_devops", private_token="xxx",
-                    jobs_to_ignore=["ignore_by_name", "folder/ignore.*"])))
-        metric = dict(type="unused_jobs", sources=sources, addition="sum")
+        self.sources["source_id"]["parameters"]["jobs_to_ignore"] = ["ignore_by_name", "folder/ignore.*"]
+        metric = dict(type="unused_jobs", sources=self.sources, addition="sum")
         response = await self.collect(
             metric,
             get_request_json_return_value=dict(
                 value=[
-                    dict(path=r"\\folder", name="pipeline", _links=dict(web=dict(href="https://azure_devops/build")),
+                    dict(path=self.path, name="pipeline", _links=dict(web=dict(href=f"{self.url}/build")),
                          latestCompletedBuild=dict(result="failed", finishTime="2019-10-15T12:24:10.1905868Z")),
-                    dict(path=r"\\folder", name="ignore_by_re", latestCompletedBuild=dict(result="failed")),
+                    dict(path=self.path, name="ignore_by_re", latestCompletedBuild=dict(result="failed")),
                     dict(path=r"\\", name="ignore_by_name", latestCompletedBuild=dict(result="failed"))
                 ]))
         expected_age = days_ago(parse("2019-10-15T12:24:10.1905868Z"))
         self.assert_measurement(
-            response, value="1",
-            api_url="https://azure_devops/_apis/build/definitions?includeLatestBuilds=true&api-version=4.1",
-            landing_url="https://azure_devops/_build",
+            response, value="1", landing_url=f"{self.url}/_build", api_url=self.api_url,
             entities=[
-                dict(name=r"folder/pipeline", key=r"folder/pipeline", url="https://azure_devops/build",
+                dict(name=self.pipeline, key=self.pipeline, url=f"{self.url}/build",
                      build_date="2019-10-15", build_age=str(expected_age), build_status="failed")])